refactor: 패키지, 프로젝트 이름 변경
