+++ extraction_module/__init__.py
@@ -1,17 +1,16 @@
 from typing import Any
-
-from structured_output_benchmark.extraction_module.base import BaseFramework
-from structured_output_benchmark.extraction_module.frameworks.openai_framework import OpenAIFramework
-from structured_output_benchmark.extraction_module.frameworks.instructor_framework import InstructorFramework
-from structured_output_benchmark.extraction_module.frameworks.langchain_tool_framework import LangchainToolFramework
-from structured_output_benchmark.extraction_module.frameworks.langchain_parser_framework import LangchainParserFramework
-from structured_output_benchmark.extraction_module.frameworks.mirascope_framework import MirascopeFramework
-from structured_output_benchmark.extraction_module.frameworks.llamaindex_framework import LlamaIndexFramework
-from structured_output_benchmark.extraction_module.frameworks.marvin_framework import MarvinFramework
-from structured_output_benchmark.extraction_module.frameworks.ollama_framework import OllamaFramework
-from structured_output_benchmark.extraction_module.frameworks.google_framework import GoogleFramework
-from structured_output_benchmark.extraction_module.frameworks.lm_format_enforcer_framework import LMFormatEnforcerFramework
-from structured_output_benchmark.extraction_module.frameworks.anthropic_framework import AnthropicFramework
+from structured_output_kit.extraction_module.base import BaseFramework
+from structured_output_kit.extraction_module.frameworks.openai_framework import OpenAIFramework
+from structured_output_kit.extraction_module.frameworks.instructor_framework import InstructorFramework
+from structured_output_kit.extraction_module.frameworks.langchain_tool_framework import LangchainToolFramework
+from structured_output_kit.extraction_module.frameworks.langchain_parser_framework import LangchainParserFramework
+from structured_output_kit.extraction_module.frameworks.mirascope_framework import MirascopeFramework
+from structured_output_kit.extraction_module.frameworks.llamaindex_framework import LlamaIndexFramework
+from structured_output_kit.extraction_module.frameworks.marvin_framework import MarvinFramework
+from structured_output_kit.extraction_module.frameworks.ollama_framework import OllamaFramework
+from structured_output_kit.extraction_module.frameworks.google_framework import GoogleFramework
+from structured_output_kit.extraction_module.frameworks.lm_format_enforcer_framework import LMFormatEnforcerFramework
+from structured_output_kit.extraction_module.frameworks.anthropic_framework import AnthropicFramework
