+++ extraction_module/frameworks/__init__.py
@@ -1,1 +1,1 @@
-"""Framework implementations for structured_output_benchmark.extraction_module"""
+"""Framework implementations for structured_output_kit.extraction_module"""
