+++ extraction_module/frameworks/ollama_framework.py
@@ -4,7 +4,7 @@ import json
 
 from ollama._client import Client
 from langfuse import observe
-from structured_output_benchmark.extraction_module.base import BaseFramework, experiment
+from structured_output_kit.extraction_module.base import BaseFramework, experiment
 
 
 class OllamaFramework(BaseFramework):
