+++ extraction_module/utils.py
@@ -2,9 +2,36 @@
 import importlib.util
 import yaml
 import os
-from typing import Dict, List, Any, Optional
+from importlib import resources
+from typing import Any
+from enum import Enum
+from dataclasses import asdict, is_dataclass
+from pydantic import BaseModel
 from loguru import logger
-from structured_output_benchmark.core.types import HostInfo
+
+def load_prompt() -> str:
+    try:
+        with resources.files("structured_output_kit").joinpath("prompt.yaml").open("r", encoding="utf-8") as f:
+            prompt_yaml = yaml.safe_load(f)
+        return prompt_yaml.get("Extract_prompt", "Extract information from the given content.")
+    except Exception:
+        if os.path.exists("prompt.yaml"):
+            with open("prompt.yaml", "r", encoding="utf-8") as f:
+                prompt_yaml = yaml.safe_load(f)
+            return prompt_yaml.get("Extract_prompt", "Extract information from the given content.")
+        return "Extract information from the given content."
+
+def response_parsing(response: Any) -> Any:
+    if isinstance(response, list):
+        response = {
+            member.value if isinstance(member, Enum) else member for member in response
+        }
+    elif is_dataclass(response):
+        response = asdict(response)
+    elif isinstance(response, BaseModel):
+        response = response.model_dump(exclude_none=True)
+    return response
+
 
 def convert_schema(schema_name: str):
     """
@@ -22,99 +49,6 @@ def convert_schema(schema_name: str):
     mod = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(mod)
     return getattr(mod, 'ExtractInfo')
-
-def extract_with_framework(
-    framework: str,
-    host_info: HostInfo,
-    content: str,
-    prompt: str,
-    schema_name: str ,
-    retries: int = 1,
-    api_delay_seconds: float = 0,
-    langfuse_trace_id: Optional[str] = None,
-    extra_kwargs: Optional[Dict[str, Any]] = None,
-    
-) -> tuple[Dict[str, Any], bool, Any]:
-    """
-    선택된 프레임워크를 사용하여 JSON 추출 수행
-    
-    Args:
-        framework: 사용할 프레임워크 이름
-        provider: LLM 제공자 (ollama, openai, google, openai_compatible)
-        model: 사용할 LLM 모델 이름
-        base_url: API 기본 URL
-        content: 추출할 텍스트 내용
-        prompt: 사용할 프롬프트
-        retries: 재시도 횟수
-        api_delay_seconds: API 호출 간 지연 시간
-    
-    Returns:
-        tuple[Dict[str, Any], bool, Any]: (추출된 데이터, 성공 여부, 지연시간(단일 값 또는 리스트))
-    """
-    try:
-        response_model = convert_schema(schema_name)
-        
-        init_kwargs = {
-            "provider": host_info.provider,
-            "model": host_info.model,
-            "base_url": host_info.base_url,
-            "api_key": host_info.api_key,
-            "prompt": prompt,
-            "response_model": response_model,
-            "api_delay_seconds": api_delay_seconds,
-            "langfuse_trace_id": langfuse_trace_id,
-            "extra_kwargs":extra_kwargs,
-        }
-        
-        # 프레임워크 인스턴스 생성
-        try:
-            # 패키지 내부 상대 임포트로 변경 (wheel 설치 환경에서도 동작)
-            from . import factory  # 순환 import 방지용 함수 내부 import
-            framework_instance = factory(framework, **init_kwargs)
-            logger.debug(f"프레임워크 {framework} 초기화 완료")
-        except Exception as e:
-            logger.error(f"프레임워크 {framework} 초기화 실패: {str(e)}")
-            return {"error": f"프레임워크 초기화 실패: {str(e)}"}, False, 0
-
-        # 입력 데이터 준비
-        inputs = {"content": content}
-        
-        # 프레임워크 실행
-        try:
-            predictions, percent_successful, latencies = framework_instance.run(
-                retries=retries,
-                inputs=inputs,
-                langfuse_trace_id=langfuse_trace_id
-            )
-            
-            logger.debug(f"프레임워크 실행 완료: 성공률 {percent_successful:.2%}, 응답 수 {len(predictions) if predictions else 0}")
-        except Exception as e:
-            logger.error(f"프레임워크 실행 중 오류: {str(e)}")
-            return {"error": f"프레임워크 실행 실패: {str(e)}"}, False, 0
-        
-        # 결과 처리
-        #if predictions and len(predictions) > 0 and not predictions[0].startswith('ERROR'):
-        if predictions and len(predictions) > 0 and isinstance(predictions[0], dict):
-            result = predictions[0]  # 첫 번째 성공한 결과 사용
-            
-            # Pydantic 모델인 경우 dict로 변환
-            if hasattr(result, 'model_dump'):
-                result = result.model_dump(exclude_none=True)
-            elif hasattr(result, 'dict'):
-                result = result.dict(exclude_none=True)
-            
-            logger.debug(f"Framework {framework} 실행 성공: 성공률 {percent_successful:.2%}")
-            return result, True, latencies
-        else:
-            if predictions[0].startswith('ERROR'):
-                logger.error(f"Framework {framework} 실행 실패: {predictions[0]}")
-            else:
-                logger.error(f"Framework {framework} 실행 실패: 성공한 응답 없음")
-            return {"error": f"성공한 응답이 없습니다: {predictions[0]}"}, False, 0
-            
-    except Exception as e:
-        logger.error(f"Framework {framework} 실행 중 예상치 못한 오류 발생: {str(e)}")
-        return {"error": str(e)}, False, 0
 
 def get_compatible_frameworks(provider: str):
     """선택한 host에 호환되는 프레임워크 목록을 반환하는 함수"""
